pyproject.toml
[tool.poetry]
name = "django-zapier-triggers"
version = "1.0.alpha2"
description = "Django (DRF) backed app for managing Zapier triggers."
license = "MIT"
authors = ["YunoJuno <code@yunojuno.com>"]
maintainers = ["YunoJuno <code@yunojuno.com>"]
readme = "README.md"
homepage = "https://github.com/yunojuno/django-zapier-trigger"
repository = "https://github.com/yunojuno/django-zapier-trigger"
documentation = "https://github.com/yunojuno/pdjango-zapier-trigger"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
]
packages = [{ include = "zapier" }]

[tool.poetry.dependencies]
python = "^3.10"
django = "^3.2 || ^4.0"
python-dateutil = "*"
pytz = "*"
requests = "*"
djangorestframework = "*"

[tool.poetry.dev-dependencies]
black = "*"
coverage = "*"
dj-database-url = "*"
factory-boy = "*"
flake8 = "*"
flake8-bandit = "*"
flake8-blind-except = "*"
flake8-docstrings = "*"
flake8-logging-format = "*"
flake8-print = "*"
freezegun = "*"
isort = "*"
mypy = "*"
pre-commit = "*"
pytest = "*"
pytest-cov = "*"
pytest-django = "*"
tox = "*"
types-freezegun = "*"
types-python-dateutil = "*"
types-pytz = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
